backend/requirements.txt
Django==4.1.2
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
factory-boy==3.2.1
Faker==15.1.3
gunicorn==20.1.0
passlib==1.7.4
psycopg2==2.9.2
pytest==7.1.3
pytest-django==4.5.2
celery==5.2.7
redis==4.3.3
openpyxl==3.0.10
